fix: make nix work out of the box on rpm-ostree
The main difference is that `restorecon` is executed against `/var/nix`
instead of against `/nix`, because that's the path that actually exists
at that point.

To make it work, SELinux policies are duplicated for the `/var/nix`
prefix.

From there onwards, no more recursive `restorecon` calls over the whole
Nix store are needed. Instead, `restorecon` is recursively called over
the system profile binaries.

Fixed a bug in the start conditions for the socket.

Start the `nix-setup.service` before `nix-daemon.service`, not before
`nix.mount`. This is because mounting the empty directory first doesn't
have any negative effect, but running the service before the mount
enters a service dependency loop that we want to avoid.

@moduon MT-83

Co-authored-by: Jairo Llopis <yajo.sk8@gmail.com>
Co-authored-by: Dusk Banks <me@bb010g.com>