Make after install hook fail on error
If something goes wrong setting up the environment, we should report
that failure out to the host package manager. The step observed to fail
was installing SELinux rules, but other steps could potentially fail,
too. Following the existing comment, we do not consider it an error if
nix-daemon fails to start.